src/embeddings-server/requirements.txt
fastapi==0.114.1
uvicorn==0.30.6
python-dotenv==1.0.1
httpx==0.27.2
haystack-ai==2.6.0
pypdf==5.0.1
markdown-it-py==3.0.0
mdit-plain==1.0.1
sentence-transformers==3.1.1
chroma-haystack==0.22.1